pyproject.toml
[tool.poetry]
name = "woogle-maps"
version = "0.1.0"
description = ""
authors = [
    "Jeffrey Lim <Jeffrey-Lim@outlook.com>",
    "Gregoire Dumont <Dumont.Gregoire.A@gmail.com>",
    "Kristóf Sándor <Emherk512@gmail.com>",
    "Daniel De Dios Allegue <DanielDeDiosAllegue@gmail.com>",
]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
    { include = "main.py" },
    { include = "gz_extract.py" },
    { include = "split_woogle_dump_per_dossier.py" },
    { include = "dashapp" },
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
seaborn = "^0.13.2"
yake = "^0.4.8"
nltk = "^3.8.1"
dash = "^2.16.0"
dash-extensions = "^1.0.12"
dash-cytoscape = "^1.0.0"
hydra-core = "^1.3.2"
tqdm = "^4.66.2"
fastparquet = "^2024.2.0"
transformers = {extras = ["torch"], version = "^4.38.2"}
sentence-transformers = "^2.5.1"
lexrank = "^0.1.0"
pandas = {extras = ["performance"], version = "^2.2.1"}
pyarrow = "^15.0.1"
epochalyst = "^0.2.0"
hdbscan = "^0.8.33"
dask = {extras = ["dataframe"], version = "^2024.3.1"}
notebook = "^7.1.2"
dash-mantine-components = "^0.12.1"
dash-iconify = "^0.1.2"
rustworkx = "^0.14.2"
pytest = "^8.1.1"
polars = "^0.20.16"
gensim = "^4.3.2"
pypdf = {extras = ["full"], version = "^4.1.0"}
reportlab = "^4.1.0"
scikit-learn = "^1.4.1.post1"
pulp = "^2.8.0"
umap-learn = "^0.5.6"
scipy = "<1.13.0"
randomname = "^0.2.1"
flask = "^3.0.3"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
pre-commit = "^3.6.2"

[tool.poetry.group.docs]

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-autodoc-typehints = "^2.0.0"
sphinxawesome-theme = "^5.1.3"
myst-parser = "^2.0.0"
pygit2 = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pydoclint]
style = "sphinx"
arg-type-hints-in-docstring = false
arg-type-hints-in-signature = true
skip-checking-short-docstrings = false
check-return-types = false
show-filenames-in-every-violation-message = true
ignore = [
    "DOC103",  # Docstring arguments are different from function arguments. Too often a false positive.
]
